The Changes for making the MySQL connection pool global and changing the connection.release in every route in the Server Route has been done.
