+++ client/app/login/password.html
@@ -1,0 +1,31 @@
+<div class="form" ng-controller='loginController'>
+  <ul class="tab-group">
+    <li class="tab active"><a href="#!/login">Login</a></li>
+    <li class="tab"><a href='#!/createAccount'>Create Account</a></li>
+  </ul>
+  <div class="tab-content"></div><!--Tab Content-->
+
+  <div id="login">
+    <h1>Recover Password!</h1><br>
+    <form ng-submit='recoverPassword()'ng-show='userName'>
+      <div class="field-wrap">
+        <label>ASURITE ID:<span class="req">*</span></label><br>
+        <input type="asurite" ng-model='forgotUserName' required autocomplete="off"/>
+      </div>
+        <button ng-disabled = "disableBtn" class="button" type='submit'>Recover</button>
+    </form>
+    <form ng-submit='retrievePassword()' ng-show='recoveryPassword'>
+      <div class="field-wrap">
+        <input type="asurite" ng-model='securityQuestion1' required autocomplete="off" disabled/>
+        <input type="text" ng-model='securityAnswer1' required autocomplete="off"/>
+      </div>
+      <div class="field-wrap">
+        <input type="text" ng-model='securityQuestion2' required autocomplete="off" disabled />
+        <input type="text" ng-model='securityAnswer2' required autocomplete="off"/>
+      </div>
+      <button ng-disabled = "disableBtn" class="button" type='submit'>Recover Password</button>
+    </form>
+    <p class="error" ng-model='userErrorMessage' ng-show='userError' style='color:red'><br><br>{{userErrorMessage}}</p>
+    <input type="hidden" ng-model = "asuID">
+  </div><!--form-->
+</div><!--Log In-->
