+++ client/app/login/passwordreset.html
@@ -1,0 +1,50 @@
+<html ng-app='app' lang="en">
+
+<head>
+    <title>TA and Grader Application System</title>
+    <!--angular dependencies-->
+    <script src="angular/angular.min.js"></script>
+    <script src="angular-route/angular-route.min.js"></script>
+    <!--module javascript source files-->
+    <script src="app/app.js"></script>
+    <script src="app/header.js"></script>
+    <script src="app/errors/errors.js"></script>
+    <script src="app/services.js"></script>
+    <script src="app/directives.js"></script>
+    <script src="app/login/login.js"></script>
+    <script src="app/account/account.js"></script>
+    <script src="app/users/users.js"></script>
+    <script src="app/programChair/programChair.js"></script>
+    <script src="app/faculty/faculty.js"></script>
+    <script src="app/student/student.js"></script>
+    <script src="app/administrative/administrative.js"></script>
+    <script src="app/application/application.js"></script>
+    <script src="app/help/help.js"></script>
+    <link rel="stylesheet" ng-href={{layout}}>
+    <link rel="stylesheet" ng-href="/css/normalize.min.css">
+</head>
+
+<body ng-show='mainDisplay' ng-model='mainDisplay'>
+  <div class="form" ng-controller='loginController'>
+    <!-- <ul class="tab-group">
+      <li class="tab active"><a href="#!/login">Login</a></li>
+      <li class="tab"><a href='#!/createAccount'>Create Account</a></li>
+    </ul> -->
+    <form ng-init = "tokenValidation()" ng-submit='changePassword()'ng-show='userName'>
+      <div class="field-wrap">
+        <label>New Password:<span class="req">*</span></label><br>
+        <input type="password" ng-model='newpassword' required autocomplete="off"/>
+      </div>
+      <div class="field-wrap">
+        <label>Confirm Password:<span class="req">*</span></label><br>
+        <input type="password" ng-model='confirmpassword' required autocomplete="off"/>
+      </div>
+        <button ng-disabled = "disableBtn" class="button" type='submit'>Change Password</button>
+    </form>
+    <p class="error" ng-model='userErrorMessage' ng-show='userError' style='color:red'><br><br>{{userErrorMessage}}</p>
+    <input type="hidden" ng-model = "asurite">
+    <div class="tab-content"></div><!--Tab Content-->
+  </div><!--Log In-->
+</body>
+
+</html>
