+++ server/serverRoutes/applicationRouters/availabilityRouter.js
@@ -5,7 +5,7 @@
 
 var express  = require('express');
 var router = express.Router();
-var mysql = require('mysql');
+//var mysql = require('mysql');
 
 // Invoked for any request passed to this router
 router.use(function(req, res, next) {
@@ -13,13 +13,15 @@ router.use(function(req, res, next) {
 });
 
 // Create mysql connection pool
-var mysql_pool  = mysql.createPool({
-	connectionLimit : 5,
-	host            : 'localhost',
-	user            : 'root',
-	password        : 'root',
-	database        : 'sblDB'
-});
+// var mysql_pool  = mysql.createPool({
+// 	connectionLimit : 5,
+// 	host            : 'localhost',
+// 	user            : 'root',
+// 	password        : 'root',
+// 	database        : 'sblDB'
+// });
+
+var mysql_pool = require('../DBConfig.js');
 
 // Deletes users old availability information and saves their new information
 router.post('/', function(req, res) {
@@ -33,24 +35,26 @@ router.post('/', function(req, res) {
 		    }
 		    connection.query('DELETE FROM Calendar WHERE ASURITE_ID = ?', [req.user.username], function(err2) {
 		        if(err2) {
+              connection.release();
 		        	console.log('Error performing query: ' + err2);
 		            throw err2;
 		        } else {
-		            connection.query('INSERT INTO Calendar (CalendarDay, StartHour, StopHour, ASURITE_ID) VALUES ?', [req.body.availableSlots], function(err3) {  
+		            connection.query('INSERT INTO Calendar (CalendarDay, StartHour, StopHour, ASURITE_ID) VALUES ?', [req.body.availableSlots], function(err3) {
 						if(err3) {
+              connection.release();
 					    	console.log('Error performing query: ' + err3);
 					        throw err3;
 					    } else {
                             connection.query('UPDATE Application SET isAvailabilityComplete = ?, AppStatus = ?, ModifiedDate = ? WHERE ASURITE_ID = ?', [req.body.isAvailabilityComplete, req.body.appStatus, dateObj, req.user.username], function(err4) {
+                              connection.release();
                                 if (err4) {
                                     throw err4;
                                 }
                                 res.sendStatus(200);
                             });
 					    }
-					    connection.release();
 				    });
-		        } 
+		        }
 		    });
 	  	});
 	} else {
@@ -62,17 +66,18 @@ router.post('/', function(req, res) {
 		    }
 		    connection.query('DELETE FROM Calendar WHERE ASURITE_ID = ?', [req.user.username], function(err2) {
 		    	if(err2) {
+            connection.release();
 		        	console.log('Error performing query: ' + err2);
 		            throw err5;
 		        } else {
                     connection.query('UPDATE Application SET isAvailabilityComplete = ?, AppStatus = ?, ModifiedDate = ? WHERE ASURITE_ID = ?',[req.body.isAvailabilityComplete, req.body.appStatus, dateObj, req.user.username], function(err3) {
+                      connection.release();
                         if (err3) {
                             throw err3;
                         }
                         res.sendStatus(200);
-                    });		        		
+                    });
 		        }
-		        connection.release();
 		    });
 		});
 	}
@@ -87,23 +92,23 @@ router.get('/', function(req, res) {
       		throw err;
     	}
     	connection.query('SELECT * FROM Calendar WHERE ASURITE_ID = ?', [req.user.username], function(err2, rows) {
+        	connection.release();
       		if(err2) {
         		console.log('Error performing query: ' + err2);
         		throw err2;
       		} else if (!rows.length) {
         		res.sendStatus(200);
-      		} else if (rows[0]) {          
+      		} else if (rows[0]) {
 				var data = [];
-				for(var i in rows) {    
-    				var item = rows[i];   
-      				data.push({ 
+				for(var i in rows) {
+    				var item = rows[i];
+      				data.push({
           				'calendarDay' 	: item.CalendarDay,
-          				'startHour'     : item.StartHour 
+          				'startHour'     : item.StartHour
       				});
     			}
         		res.send({'data' : data});
       		}
-      		connection.release();
     	});
   	});
 });
