+++ server/serverRoutes/applicationRouters/coursesRouter.js
@@ -5,7 +5,7 @@
 
 var express  = require('express');
 var router = express.Router();
-var mysql = require('mysql');
+//var mysql = require('mysql');
 
 // Invoked for any request passed to this router
 router.use(function(req, res, next) {
@@ -13,16 +13,18 @@ router.use(function(req, res, next) {
 });
 
 //  Create mysql connection pool
-var mysql_pool  = mysql.createPool({
-    connectionLimit : 5,
-    host            : 'localhost',
-    user            : 'root',
-    password        : 'root',
-    database        : 'sblDB'
-});
+// var mysql_pool  = mysql.createPool({
+//     connectionLimit : 5,
+//     host            : 'localhost',
+//     user            : 'root',
+//     password        : 'root',
+//     database        : 'sblDB'
+// });
+
+var mysql_pool = require('../DBConfig.js');
 
 // Creates/Updates user course choices
-router.post('/', function(req, res) { 
+router.post('/', function(req, res) {
     var dateObj = new Date().toISOString().slice(0, 19).replace('T', ' ');
     mysql_pool.getConnection(function(err, connection) {
         if (err) {
@@ -33,33 +35,36 @@ router.post('/', function(req, res) {
         if (req.body.data[0].length === 0) {
             connection.query('DELETE FROM Course_Competencies WHERE ASURITE_ID = ?', [req.user.username], function(err2) {
                 if(err2) {
+                  connection.release();
                     console.log('Error performing query: ' + err2);
                     throw err2;
                 }
             });
-        } 
+        }
         if (req.body.data[0].length != 0) {
             connection.query('DELETE FROM Course_Competencies WHERE ASURITE_ID = ?', [req.user.username], function(err3) {
                 if(err3) {
+                  connection.release();
                     console.log('Error performing query: ' + err3);
                     throw err3;
                 } else {
-                    connection.query('INSERT INTO Course_Competencies (isCourse, isPrefer, isQualified, isPreviouslyTA, isPreviouslyGrader, OtherCourse, isOtherPrefer, isOtherQualified, isOtherPreviouslyTA, isOtherPreviouslyGrader, ASURITE_ID) VALUES ?', [req.body.data[0]], function(err4) { 
+                    connection.query('INSERT INTO Course_Competencies (isCourse, isPrefer, isQualified, isPreviouslyTA, isPreviouslyGrader, OtherCourse, isOtherPrefer, isOtherQualified, isOtherPreviouslyTA, isOtherPreviouslyGrader, ASURITE_ID) VALUES ?', [req.body.data[0]], function(err4) {
                         if(err4) {
+                          connection.release();
                             console.log('Error performing query: ' + err4);
                             throw err4;
                         }
                     });
-                } 
+                }
             });
         }
         connection.query('UPDATE Application SET isCoursesComplete = ?, AppStatus = ?, ModifiedDate = ?, DateSubmitted = ?  WHERE ASURITE_ID = ?', [req.body.isCoursesComplete, req.body.appStatus, dateObj, dateObj, req.user.username], function(err5) {
+          connection.release();
             if (err5) {
                 throw err5;
             }
             res.sendStatus(200);
         });
-        connection.release();
     });
 });
 
@@ -72,23 +77,23 @@ router.get('/', function(req, res) {
             console.log('Error getting mysql_pool connection: ' + err);
             throw err;
         }
-        connection.query('SELECT isCourse, isPrefer, isQualified, isPreviouslyTA, isPreviouslyGrader, OtherCourse, isOtherPrefer, isOtherQualified, isOtherPreviouslyTA, isOtherPreviouslyGrader FROM Course_Competencies WHERE ASURITE_ID = ?', [req.user.username], function(err2, rows) { 
+        connection.query('SELECT isCourse, isPrefer, isQualified, isPreviouslyTA, isPreviouslyGrader, OtherCourse, isOtherPrefer, isOtherQualified, isOtherPreviouslyTA, isOtherPreviouslyGrader FROM Course_Competencies WHERE ASURITE_ID = ?', [req.user.username], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
             } else if (rows[0]) {
-                for (var i = 0; i < rows.length; i++) {                 
+                for (var i = 0; i < rows.length; i++) {
                     courses.push(rows[i]);
                 }
-            }   
-            connection.release();
+            }
             sendPopulateResponse(res, courses);
         });
     });
 });
 
 function sendPopulateResponse (res, courses) {
-    res.send({'data':{'courseData': courses}});      
+    res.send({'data':{'courseData': courses}});
 }
 
 // Gets Courses from Courses Table
@@ -99,7 +104,8 @@ router.get('/courseTable', function(req, res) {
             console.log('Error getting mysql_pool connection: ' + err);
             throw err;
         }
-        connection.query('SELECT CourseSection, CourseName FROM Courses', function(err2, rows) { 
+        connection.query('SELECT CourseSection, CourseName FROM Courses', function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -108,9 +114,9 @@ router.get('/courseTable', function(req, res) {
                 for (var i = 0; i < rows.length; i++) {
                      var Courses = {
                     'CourseSection' : rows[i].CourseSection,
-                    'CourseName'    : rows[i].CourseName}; 
+                    'CourseName'    : rows[i].CourseName};
                     CourseData.push(Courses);
-                }   
+                }
                 var sortByProperty = function (property) {
                     return function (x, y) {
                         return ((x[property].substring(4,7) === y[property].substring(4,7)) ? 0 : ((x[property].substring(4,7) > y[property].substring(4,7)) ? 1 : -1));
@@ -119,8 +125,7 @@ router.get('/courseTable', function(req, res) {
                 CourseData.sort(sortByProperty('CourseSection'));
                 var response = JSON.stringify(CourseData);
                 res.send(response);
-            }   
-            connection.release();
+            }
         });
     });
 });
