+++ server/serverRoutes/applicationRouters/educationRouter.js
@@ -5,7 +5,7 @@
 
 var express  = require('express');
 var router = express.Router();
-var mysql = require('mysql');
+//var mysql = require('mysql');
 
 // Invoked for any request passed to this router
 router.use(function(req, res, next) {
@@ -13,13 +13,15 @@ router.use(function(req, res, next) {
 });
 
 //  Create mysql connection pool
-var mysql_pool  = mysql.createPool({
-    connectionLimit : 5,
-    host            : 'localhost',
-    user            : 'root',
-    password        : 'root',
-    database        : 'sblDB'
-});
+// var mysql_pool  = mysql.createPool({
+//     connectionLimit : 5,
+//     host            : 'localhost',
+//     user            : 'root',
+//     password        : 'root',
+//     database        : 'sblDB'
+// });
+
+var mysql_pool = require('../DBConfig.js');
 
 // Checks if user already has this information saved then Saves/Updates user entered information into application
 router.post('/', function(req, res) {
@@ -29,13 +31,15 @@ router.post('/', function(req, res) {
             console.log('Error getting mysql_pool connection: ' + err);
             throw err;
         }
-        connection.query('SELECT * FROM Application WHERE ASURITE_ID = ?', [req.body.ASURITE_ID], function(err2, rows) { 
+        connection.query('SELECT * FROM Application WHERE ASURITE_ID = ?', [req.body.ASURITE_ID], function(err2, rows) {
             if(err2) {
+              connection.release();
                 console.log('Error performing query: ' + err2);
                 throw err2;
             } else if (!rows.length) {
                 connection.query('INSERT INTO Application SET ?', [req.body], function(err3) {
                     if(err3) {
+                      connection.release();
                         console.log('Error performing query: ' + err3);
                         throw err3;
                     } else {
@@ -44,15 +48,15 @@ router.post('/', function(req, res) {
                 });
             } else if (rows[0]) {
                 connection.query('UPDATE Application SET ? WHERE ASURITE_ID = ?', [req.body, req.body.ASURITE_ID], function(err4) {
+                  connection.release();
                     if(err4) {
                         console.log('Error performing query: ' + err4);
                         throw err4;
                     } else {
                         res.sendStatus(200);
                     }
-                }); 
+                });
             }
-            connection.release();
         });
     });
 });
@@ -66,16 +70,16 @@ router.get('/', function(req, res) {
             throw err;
         }
         connection.query('SELECT Application.EducationLevel, Application.GPA, Application.DegreeProgram, Application.isAcademicProbation, Application.isFourPlusOne, Application.isFullTime, Application.FirstSession, Application.GraduationDate, Attachment.IposName, Attachment.TranscriptName FROM Application LEFT JOIN Attachment ON Application.ASURITE_ID = Attachment.ASURITE_ID WHERE Application.ASURITE_ID = ?', [req.user.username], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
             } else if (!rows.length) {
                 res.sendStatus(200);
             } else if (rows[0]) {
-                res.send({'EducationLevel' : rows[0].EducationLevel, 'GPA' : rows[0].GPA, 'DegreeProgram' : rows[0].DegreeProgram, 'isAcademicProbation' : rows[0].isAcademicProbation, 
+                res.send({'EducationLevel' : rows[0].EducationLevel, 'GPA' : rows[0].GPA, 'DegreeProgram' : rows[0].DegreeProgram, 'isAcademicProbation' : rows[0].isAcademicProbation,
                           'isFourPlusOne' : rows[0].isFourPlusOne, 'isFullTime' : rows[0].isFullTime, 'FirstSession' : rows[0].FirstSession, 'GraduationDatel' : rows[0].GraduationDate, 'ipos' : rows[0].IposName, 'transcript' : rows[0].TranscriptName});
-            } 
-            connection.release();
+            }
         });
     });
 });
