+++ server/serverRoutes/applicationRouters/educationTranscriptUploadRouter.js
@@ -5,7 +5,7 @@
 
 var express  = require('express');
 var router = express.Router();
-var mysql = require('mysql');
+//var mysql = require('mysql');
 var fs = require("fs");
 var multer  = require('multer');
 
@@ -39,13 +39,15 @@ var storage = multer.diskStorage({
 router.use(multer({storage : storage}).any());
 
 //  Create mysql connection pool
-var mysql_pool  = mysql.createPool({
-    connectionLimit : 5,
-    host            : 'localhost',
-    user            : 'root',
-    password        : 'root',
-    database        : 'sblDB'
-});
+// var mysql_pool  = mysql.createPool({
+//     connectionLimit : 5,
+//     host            : 'localhost',
+//     user            : 'root',
+//     password        : 'root',
+//     database        : 'sblDB'
+// });
+
+var mysql_pool = require('../DBConfig.js');
 
 // Save transcript attachment information in database
 router.post('/', function(req, res) {
@@ -60,17 +62,20 @@ router.post('/', function(req, res) {
             } else {
                 connection.query('SELECT TranscriptName FROM Attachment WHERE ASURITE_ID = ?', [req.user.username], function(err2, rows) {
                     if(err2) {
+                      connection.release();
                         console.log('Error performing query: ' + err2);
                         throw err2;
                     } else if (!rows.length) {
                         connection.query('INSERT INTO Attachment (TranscriptName, TranscriptUploadDate, ASURITE_ID) VALUES (?, ?, ?)', [req.files[0].originalname, uploadTime, req.user.username], function(err3) {
                             if(err3) {
+                              connection.release();
                                 console.log('Error performing query: ' + err3);
                                 throw err3;
                             }
                         });
                     } else {
                         connection.query('UPDATE Attachment SET TranscriptName = ?, TranscriptUploadDate = ? WHERE ASURITE_ID = ?', [req.files[0].originalname, uploadTime, req.user.username], function(err4) {
+                          connection.release();
                             if(err4) {
                                 console.log('Error performing query: ' + err4);
                                 throw err4;
@@ -79,7 +84,6 @@ router.post('/', function(req, res) {
                     }
                 });
             }
-        connection.release();
         });
     }
     res.sendStatus(200);
