+++ server/serverRoutes/facultyRouter/facultyRouter.js
@@ -5,7 +5,7 @@
 
 var express  = require('express');
 var router = express.Router();
-var mysql = require('mysql');
+//var mysql = require('mysql');
 var fs = require('fs');
 var bcrypt = require('bcrypt');
 
@@ -15,13 +15,15 @@ router.use(function(req, res, next) {
 });
 
 //  Create mysql connection pool
-var mysql_pool  = mysql.createPool({
-    connectionLimit : 5,
-    host            : 'localhost',
-    user            : 'root',
-    password        : 'root',
-    database        : 'sblDB'
-});
+// var mysql_pool  = mysql.createPool({
+//     connectionLimit : 5,
+//     host            : 'localhost',
+//     user            : 'root',
+//     password        : 'root',
+//     database        : 'sblDB'
+// });
+
+var mysql_pool = require('../DBConfig.js');
 
 // Get Student Application Names
 router.get('/evaluations/appnames', function(req, res) {
@@ -32,6 +34,7 @@ router.get('/evaluations/appnames', function(req, res) {
             throw err;
         }
         connection.query("SELECT CONCAT(FirstName, ' ', LastName) AS StudentName, ASURITE_ID FROM User_ WHERE UserRole = 'student'", function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -50,7 +53,6 @@ router.get('/evaluations/appnames', function(req, res) {
                 var response = JSON.stringify(Names);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -64,6 +66,7 @@ router.post('/evaluation', function(req, res) {
             throw err;
         }
         connection.query('INSERT INTO Student_Evaluation SET ?', [req.body], function(err2) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -71,7 +74,6 @@ router.post('/evaluation', function(req, res) {
                 res.sendStatus(200);
                 }
         });
-        connection.release();
     });
 });
 
@@ -84,6 +86,7 @@ router.post('/applicationNames', function(req, res) {
             throw err;
         }
         connection.query("SELECT concat(FirstName, ' ', LastName) Name, User_.ASURITE_ID From User_ INNER JOIN Application ON User_.ASURITE_ID = Application.ASURITE_ID WHERE AppStatus = 'complete'", [req.body.course], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -100,7 +103,6 @@ router.post('/applicationNames', function(req, res) {
                 var response = JSON.stringify(Students);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -114,6 +116,7 @@ router.post('/applications', function(req, res) {
             throw err;
         }
         connection.query("SELECT concat(FirstName, ' ', LastName) Name, isQualified, isPrefer, isPreviouslyTA, isPreviouslyGrader, GPA, User_.ASURITE_ID From User_ INNER JOIN Course_Competencies ON User_.ASURITE_ID = Course_Competencies.ASURITE_ID INNER JOIN Application ON User_.ASURITE_ID = Application.ASURITE_ID WHERE isCourse = ? AND (isAcademicProbation = 0 OR isAcademicProbation IS NULL) AND AppStatus = 'complete' AND (SpeakTest >= 26 OR SpeakTest IS NULL) AND ((isQualified = 1 OR isPrefer = 1) OR (isPreviouslyGrader = 1 OR isPreviouslyTA = 1)) ORDER BY isQualified DESC, isPrefer DESC, isPreviouslyTA DESC, isPreviouslyGrader DESC, GPA DESC", [req.body.course], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -135,7 +138,6 @@ router.post('/applications', function(req, res) {
                 var response = JSON.stringify(StudentNames);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -149,6 +151,7 @@ router.post('/applicationTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT CONCAT(User_.FirstName, ' ', User_.LastName) AS Name, EducationLevel, DegreeProgram, isFourPlusOne, GPA, isInternationalStudent, SpeakTest, DATE_FORMAT(FirstSession, '%m/%d/%Y') AS FirstSession, GraduationDate, TimeCommitment, isTA, isGrader, CurrentEmployer, WorkHours, isWorkedASU FROM Application INNER JOIN User_ ON Application.ASURITE_ID = User_.ASURITE_ID WHERE Application.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -175,7 +178,6 @@ router.post('/applicationTable', function(req, res) {
                 var response = JSON.stringify(AppInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -189,6 +191,7 @@ router.post('/languagesTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT isLanguage, LanguageLevel, OtherLanguage, OtherLevel FROM Languages INNER JOIN User_ ON Languages.ASURITE_ID = User_.ASURITE_ID WHERE Languages.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -207,7 +210,6 @@ router.post('/languagesTable', function(req, res) {
                 var response = JSON.stringify(LanguageInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -221,6 +223,7 @@ router.post('/ideTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT isIDE, OtherIDE FROM IDEs INNER JOIN User_ ON IDEs.ASURITE_ID = User_.ASURITE_ID WHERE IDEs.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -237,7 +240,6 @@ router.post('/ideTable', function(req, res) {
                 var response = JSON.stringify(IDEInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -251,6 +253,7 @@ router.post('/toolsTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT isTool, OtherTool FROM Collaborative_Tools INNER JOIN User_ ON Collaborative_Tools.ASURITE_ID = User_.ASURITE_ID WHERE Collaborative_Tools.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -267,7 +270,6 @@ router.post('/toolsTable', function(req, res) {
                 var response = JSON.stringify(ToolInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -281,6 +283,7 @@ router.post('/coursesTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT isCourse, isPrefer, isQualified, isPreviouslyTA, isPreviouslyGrader, OtherCourse, isOtherPrefer, isOtherQualified, isOtherPreviouslyTA, isOtherPreviouslyGrader FROM Course_Competencies INNER JOIN User_ ON Course_Competencies.ASURITE_ID = User_.ASURITE_ID WHERE Course_Competencies.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -305,7 +308,6 @@ router.post('/coursesTable', function(req, res) {
                 var response = JSON.stringify(CourseInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -319,6 +321,7 @@ router.post('/calendarTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT CalendarDay, TIME_FORMAT(StartHour, '%h:%i %p') AS StartHour, TIME_FORMAT(StopHour, '%h:%i %p') AS StopHour FROM Calendar INNER JOIN User_ ON Calendar.ASURITE_ID = User_.ASURITE_ID WHERE Calendar.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -336,7 +339,6 @@ router.post('/calendarTable', function(req, res) {
                 var response = JSON.stringify(CalendarInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -350,6 +352,7 @@ router.post('/attachmentTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT IposName, TranscriptName, ResumeName, DATE_FORMAT(IposUploadDate, '%m/%d/%Y') AS IposUploadDate, DATE_FORMAT(TranscriptUploadDate, '%m/%d/%Y') AS TranscriptUploadDate, DATE_FORMAT(ResumeUploadDate, '%m/%d/%Y') AS ResumeUploadDate, Attachment.ASURITE_ID FROM Attachment INNER JOIN User_ ON Attachment.ASURITE_ID = User_.ASURITE_ID WHERE Attachment.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -368,7 +371,6 @@ router.post('/attachmentTable', function(req, res) {
                 var response = JSON.stringify(AttachmentInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -409,6 +411,7 @@ router.post('/evaluationTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT DATE_FORMAT(DateCreated, '%m/%d/%Y') AS DateCreated, QOneScore, QOneComments, QTwoScore, QTwoComments, QThreeScore, QThreeComments, QFourScore, QFourComments FROM Student_Evaluation WHERE Student_Evaluation.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -432,7 +435,6 @@ router.post('/evaluationTable', function(req, res) {
                 var response = JSON.stringify(EvalInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -446,6 +448,7 @@ router.get('/courses', function(req, res) {
             throw err;
         }
         connection.query("SELECT CourseSection FROM Courses", function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -467,7 +470,6 @@ router.get('/courses', function(req, res) {
                 var response = JSON.stringify(CourseInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -482,6 +484,7 @@ router.post('/facultySetUserPassword', function(req, res) {
             throw err;
         }
         connection.query('SELECT UserPassword FROM User_ WHERE ASURITE_ID = ?', [req.body.ASURITE_ID], function(err2, rows){
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -490,7 +493,6 @@ router.post('/facultySetUserPassword', function(req, res) {
             } else {
                 res.send({'error' : 1}); // Responds error 1 if user not found
             }
-          connection.release();
         });
     });
 });
@@ -514,13 +516,13 @@ function changePassword(response, req, validation) {
                         throw err;
                     }
                     connection.query('UPDATE User_ SET UserPassword = ? WHERE ASURITE_ID = ?', [req.body.NewPassword, req.body.ASURITE_ID], function(err2, rows) {
+                      connection.release();
                         if(err2) {
                             console.log('Error performing query: ' + err2);
                             throw err2;
                         } else {
                             response.sendStatus(200);
                         }
-                        connection.release();
                     });
                 });
             }
@@ -538,6 +540,7 @@ router.post('/facultyGetClassInfo', function(req, res) {
             throw err;
         }
         connection.query("SELECT * FROM Schedule_ WHERE concat(FirstName, ' ', LastName) = ?", [req.body.facultyName], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query2: ' + err2);
                 throw err2;
@@ -546,7 +549,6 @@ router.post('/facultyGetClassInfo', function(req, res) {
             } else if (rows[0]) {
                 res.send(rows);
             }
-            connection.release();
         });
     });
 });
@@ -559,6 +561,7 @@ router.post('/facultyGetRequests', function(req, res) {
             throw err;
         }
         connection.query("SELECT * FROM Student_Request WHERE ScheduleID = ?", [req.body.scheduleID], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query2: ' + err2);
                 throw err2;
@@ -567,7 +570,6 @@ router.post('/facultyGetRequests', function(req, res) {
             } else if (rows[0]) {
                 res.send(rows);
             }
-            connection.release();
         });
     });
 });
@@ -582,6 +584,7 @@ router.post('/facultyGetAssignedStudents', function(req, res) {
             throw err;
         }
         connection.query("SELECT * FROM Placement WHERE ScheduleID = ?", [req.body.scheduleID], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query2: ' + err2);
                 throw err2;
@@ -590,7 +593,6 @@ router.post('/facultyGetAssignedStudents', function(req, res) {
             } else if (rows[0]) {
                 res.send(rows);
             }
-            connection.release();
         });
     });
 });
@@ -614,17 +616,20 @@ router.post('/makeRequests', function(req, res) {
         }
         connection.query('SELECT * FROM User_ WHERE ASURITE_ID = ?', [student], function(err2, rows) {
             if(err2) {
+              connection.release();
                 console.log('Error performing query: ' + err2);
                 throw err2;
             } else if (rows[0]) {
                 connection.query('SELECT * FROM Student_Request WHERE ScheduleID = ?', [req.body.ScheduleID], function(err3, rows) {
-                    if(err2) {
+                    if(err3) {
+                      connection.release();
                         console.log('Error performing query 1: ' + err3);
                         throw err3;
                     } else if (rows[0]) {
                         if (req.body.studentOne) {
                             connection.query('UPDATE Student_Request SET Rank1 = ? WHERE ScheduleID = ?', [student, req.body.ScheduleID], function(err4) {
                                 if(err4) {
+                                  connection.release();
                                     console.log('Error performing query 2 : ' + err4);
                                     throw err4;
                                 } else {
@@ -634,6 +639,7 @@ router.post('/makeRequests', function(req, res) {
                         } else if (req.body.studentTwo) {
                             connection.query('UPDATE Student_Request SET Rank2 = ? WHERE ScheduleID = ?', [student, req.body.ScheduleID], function(err5) {
                                 if(err5) {
+                                  connection.release();
                                     console.log('Error performing query 3: ' + err5);
                                     throw err5;
                                 } else {
@@ -645,6 +651,7 @@ router.post('/makeRequests', function(req, res) {
                         if (req.body.studentOne) {
                             connection.query('INSERT INTO Student_Request (Rank1 ,ScheduleID) VALUES (?,?)', [student, req.body.ScheduleID], function(err6) {
                                 if(err6) {
+                                  connection.release();
                                     console.log('Error performing query 4: ' + err6);
                                     throw err6;
                                 } else {
@@ -654,6 +661,7 @@ router.post('/makeRequests', function(req, res) {
                         } else if (req.body.studentTwo) {
                             connection.query('INSERT INTO Student_Request (Rank2 ,ScheduleID) VALUES (?,?)', [student, req.body.ScheduleID], function(err7) {
                                 if(err7) {
+                                  connection.release();
                                     console.log('Error performing query 5: ' + err7);
                                     throw err7;
                                 } else {
@@ -671,12 +679,14 @@ router.post('/makeRequests', function(req, res) {
                 }
                 connection.query('SELECT * FROM Student_Request WHERE ScheduleID = ?', [req.body.ScheduleID], function(err8, rows) {
                     if(err8) {
+                      connection.release();
                         console.log('Error performing query: ' + err8);
                         throw err8;
                     } else if (rows[0]) {
                         if (req.body.studentOne === null) {
                             connection.query('UPDATE Student_Request SET Rank1 = ? WHERE ScheduleID = ?', [null, req.body.ScheduleID], function(err9) {
                                 if(err9) {
+                                  connection.release();
                                     console.log('Error performing query: ' + err9);
                                     throw err9;
                                 } else {
@@ -686,7 +696,9 @@ router.post('/makeRequests', function(req, res) {
                             });
                         } else if (req.body.studentTwo === null) {
                             connection.query('UPDATE Student_Request SET Rank2 = ? WHERE ScheduleID = ?', [null, req.body.ScheduleID], function(err10) {
+                              connection.release();
                                 if(err10) {
+                                  connection.release();
                                     console.log('Error performing query: ' + err10);
                                     throw err10;
                                 } else {
@@ -705,7 +717,6 @@ router.post('/makeRequests', function(req, res) {
                 });
             }
         });
-        connection.release();
     });
 
 });
@@ -719,6 +730,7 @@ router.post('/editRequests', function(req, res) {
             throw err;
         } else {
             connection.query('UPDATE Student_Request SET Rank1 = ?, Rank2 = ? WHERE ScheduleID = ?', [req.body.studentOne, null, req.body.ScheduleID], function(err2) {
+              connection.release();
                 if(err2) {
                     console.log('Error performing query: ' + err2);
                     throw err2;
@@ -727,7 +739,6 @@ router.post('/editRequests', function(req, res) {
                 }
             });
         }
-        connection.release();
     });
 });
