+++ server/serverRoutes/homeRouters/studentActionsRouter.js
@@ -5,7 +5,7 @@
 
 var express  = require('express');
 var router = express.Router();
-var mysql = require('mysql');
+//var mysql = require('mysql');
 
 // Invoked for any request passed to this router
 router.use(function(req, res, next) {
@@ -13,13 +13,15 @@ router.use(function(req, res, next) {
 });
 
 //  Create mysql connection pool
-var mysql_pool  = mysql.createPool({
-    connectionLimit : 5,
-    host            : 'localhost',
-    user            : 'root',
-    password        : 'root',
-    database        : 'sblDB'
-});
+// var mysql_pool  = mysql.createPool({
+//     connectionLimit : 5,
+//     host            : 'localhost',
+//     user            : 'root',
+//     password        : 'root',
+//     database        : 'sblDB'
+// });
+
+var mysql_pool = require('../DBConfig.js');
 
 router.post('/getStudentActions', function(req, res) {
     mysql_pool.getConnection(function(err, connection) {
@@ -30,16 +32,17 @@ router.post('/getStudentActions', function(req, res) {
         }
         connection.query('SELECT AppStatus FROM Application WHERE ASURITE_ID = ?', [req.body.user], function(err2, rows) {
             if(err2) {
+              connection.release();
                 console.log('Error performing query: ' + err2);
                 throw err2;
             }
             var deadline = 'SELECT CurrentSemester, DeadlineDate FROM Deadline';
             connection.query(deadline, function(err11, deadlineRow) {
-                var deadlineInfo = {deadlineSemester:deadlineRow[0].CurrentSemester, deadlineDate:deadlineRow[0].DeadlineDate};     
-                if (rows[0].AppStatus === 'new') {                
-                    res.send({hasAppActions:0, onProbation:0, deadlineSemester:deadlineRow[0].CurrentSemester, deadlineDate:deadlineRow[0].DeadlineDate});          
+                var deadlineInfo = {deadlineSemester:deadlineRow[0].CurrentSemester, deadlineDate:deadlineRow[0].DeadlineDate};
+                if (rows[0].AppStatus === 'new') {
+                    res.send({hasAppActions:0, onProbation:0, deadlineSemester:deadlineRow[0].CurrentSemester, deadlineDate:deadlineRow[0].DeadlineDate});
                 } else {
-                    var missingPages = [];               
+                    var missingPages = [];
                     var contactQuery = 'SELECT PhoneNumber, MobileNumber, AddressOne, AddressCountry, AddressCity, AddressState, AddressZip FROM Application WHERE ASURITE_ID = ?';
                     var educationQuery = 'SELECT EducationLevel, GPA, DegreeProgram, isAcademicProbation, isFourPlusOne, FirstSession, GraduationDate FROM Application WHERE ASURITE_ID = ?';
                     var employmentQuery = 'SELECT TimeCommitment, isTA, isGrader, CurrentEmployer, WorkHours, isInternationalStudent, SpeakTest FROM Application WHERE ASURITE_ID = ?';
@@ -48,6 +51,7 @@ router.post('/getStudentActions', function(req, res) {
                     var toolQuery = 'SELECT ToolID FROM Collaborative_Tools WHERE ASURITE_ID = ?';
                     connection.query('SELECT isContactComplete, isEducationComplete, isEmploymentComplete, isAvailabilityComplete, isLanguagesComplete, isCoursesComplete FROM Application WHERE ASURITE_ID = ?', [req.body.user], function(err3, top) {
                         if (err3) {
+                          connection.release();
                             throw err3;
                         } else {
                             // makes function calls to check page fields and returns missing items and probation status
@@ -60,23 +64,24 @@ router.post('/getStudentActions', function(req, res) {
                                         missingPages.push(data);
                                     }
                                     employmentCheck(function(data) {
-                                        if (!!data) {                                      
+                                        if (!!data) {
                                             missingPages.push(data);
                                         }
                                         availabilityCheck(function(data) {
-                                            if (!!data) {                                      
+                                            if (!!data) {
                                                 missingPages.push(data);
                                             }
                                             languagesCheck(function(data) {
-                                                if (!!data) {                                      
+                                                if (!!data) {
                                                     missingPages.push(data);
                                                 }
                                                 coursesCheck(function(data) {
-                                                    if (!!data) {                                      
+                                                    if (!!data) {
                                                         missingPages.push(data);
                                                     }
                                                     connection.query('SELECT isAcademicProbation FROM Application WHERE ASURITE_ID = ?', [req.body.user], function(err4, rows) {
                                                         if (err4) {
+                                                          connection.release();
                                                             throw err4;
                                                         }
                                                         if (!rows[0].isAcademicProbation) {
@@ -86,22 +91,23 @@ router.post('/getStudentActions', function(req, res) {
                                                             res.send({hasAppActions:1, appActions:missingPages, pageStatuses:top , onProbation:rows[0].isAcademicProbation, deadlineSemester:deadlineRow[0].CurrentSemester, deadlineDate:deadlineRow[0].DeadlineDate});
                                                         } else {
                                                             res.send({hasAppActions:2, onProbation:rows[0].isAcademicProbation,deadlineSemester:deadlineRow[0].CurrentSemester, deadlineDate:deadlineRow[0].DeadlineDate});
-                                                        }                                                   
-                                                    });                                               
+                                                        }
+                                                    });
                                                 });
                                             });
                                         });
                                     });
                                 });
-                            });                        
+                            });
 
                             // checks if contact page is complete and if not, check missing fields
                             function contactCheck(callback) {
-                                if (top[0].isContactComplete === 0 || top[0].isContactComplete === null) { 
+                                if (top[0].isContactComplete === 0 || top[0].isContactComplete === null) {
                                     connection.query(contactQuery, [req.body.user], function(err4, rows) {
                                         if (err4) {
+                                          connection.release();
                                             throw err4;
-                                        }                                                                    
+                                        }
                                         callback({page:'Contact', missingItems:checkContactInfo(rows[0])});
                                     });
                                 } else {
@@ -114,8 +120,9 @@ router.post('/getStudentActions', function(req, res) {
                                 if (top[0].isEducationComplete === 0 || top[0].isEducationComplete === null) {
                                     connection.query(educationQuery, [req.body.user], function(err5, rows) {
                                         if (err5) {
+                                          connection.release();
                                             throw err5;
-                                        }     
+                                        }
                                         callback({page:'Education', missingItems:checkEducationInfo(rows[0])});
                                     });
                                 } else {
@@ -128,8 +135,9 @@ router.post('/getStudentActions', function(req, res) {
                                 if (top[0].isEmploymentComplete === 0 || top[0].isEmploymentComplete === null) {
                                     connection.query(employmentQuery, [req.body.user], function(err6, rows) {
                                         if (err6) {
+                                          connection.release();
                                             throw err6;
-                                        }                                                                    
+                                        }
                                         callback({page:'Employment', missingItems:checkEmploymentInfo(rows[0])});
                                     });
                                 } else {
@@ -152,6 +160,7 @@ router.post('/getStudentActions', function(req, res) {
                                 if (top[0].isLanguagesComplete === 0 || top[0].isLanguagesComplete === null) {
                                     connection.query(languageQuery, [req.body.user], function(err7, rows) {
                                         if (err7) {
+                                          connection.release();
                                             throw err7;
                                         }
                                         if (rows.length === 0) {
@@ -159,12 +168,14 @@ router.post('/getStudentActions', function(req, res) {
                                         }
                                         connection.query(ideQuery, [req.body.user], function(err8, rows) {
                                             if (err8) {
+                                              connection.release();
                                                 throw err8;
                                             }
                                             if (rows.length === 0) {
                                                 langItems.push('Select or add at least one IDE');
                                             }
                                             connection.query(toolQuery, [req.body.user], function(err9, rows) {
+                                              connection.release();
                                                 if (err9) {
                                                     throw err9;
                                                 }
@@ -174,7 +185,7 @@ router.post('/getStudentActions', function(req, res) {
                                                 callback({page:'Languages', missingItems:langItems});
                                             });
                                         });
-                                    });           
+                                    });
                                 } else {
                                     callback(false);
                                 }
@@ -188,7 +199,7 @@ router.post('/getStudentActions', function(req, res) {
                                     callback(false);
                                 }
                             }
-                        } // end else   
+                        } // end else
                         // checks for missing items for Contact Info page
                         function checkContactInfo(row) {
                             var items = [];
@@ -284,12 +295,11 @@ router.post('/getStudentActions', function(req, res) {
                                 }
                                 else {
                                 return ({CurrentSemester:rows[0].CurrentSemester, DeadlineDate:rows[0].DeadlineDate});
-                                }    
+                                }
                             });
                         }
-                    });                       
-                } 
-                connection.release();
+                    });
+                }
             });
         });
     });
@@ -303,15 +313,15 @@ router.post('/getApplicationFeedback', function(req, res) {
             throw err;
         }
         connection.query("SELECT Feedback.FeedbackText, Feedback.hasComplied FROM Feedback LEFT JOIN Application ON Feedback.AppID = Application.AppID WHERE ASURITE_ID = ?", [req.body.user], function(err2, rows) {
+          connection.release();
             if(err2) {
                     console.log('Error performing query: ' + err2);
                     throw err2;
-            } 
+            }
             if (rows[0]) {
-                connection.release();
+                //connection.release();
                 res.send({feedback:rows[0].FeedbackText, hasComplied:rows[0].hasComplied});
             } else {
-                connection.release();
                 res.send({feedback:'No feedback from Program Chair'});
             }
         });
@@ -326,11 +336,11 @@ router.post('/saveFeedbackHasComplied', function(req, res) {
             throw err;
         }
         connection.query("UPDATE Feedback LEFT JOIN Application ON Feedback.AppID = Application.AppID SET Feedback.hasComplied = ? WHERE ASURITE_ID = ?", [req.body.hasComplied, req.body.id], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
-            } 
-            connection.release();
+            }
             res.sendStatus(200);
         });
     });
