+++ server/serverRoutes/programChairRouters/programChairRouter.js
@@ -5,7 +5,7 @@
 
 var express  = require('express');
 var router = express.Router();
-var mysql = require('mysql');
+//var mysql = require('mysql');
 var fs = require('fs');
 var bcrypt = require('bcrypt');
 
@@ -15,13 +15,15 @@ router.use(function(req, res, next) {
 });
 
 //  Create mysql connection pool
-var mysql_pool  = mysql.createPool({
-    connectionLimit : 5,
-    host            : 'localhost',
-    user            : 'root',
-    password        : 'root',
-    database        : 'sblDB'
-});
+// var mysql_pool  = mysql.createPool({
+//     connectionLimit : 5,
+//     host            : 'localhost',
+//     user            : 'root',
+//     password        : 'root',
+//     database        : 'sblDB'
+// });
+
+var mysql_pool = require('../DBConfig.js');
 
 // Returns data to populate class drop down on program chair home page
 router.get('/getClassNames', function(req, res) {
@@ -32,6 +34,7 @@ router.get('/getClassNames', function(req, res) {
             throw err;
         } else {
             connection.query('SELECT * FROM Schedule_', function(err2, rows) {
+              connection.release();
                 if(err2) {
                     console.log('Error performing query: ' + err2);
                     throw err2;
@@ -40,7 +43,6 @@ router.get('/getClassNames', function(req, res) {
                 } else if (rows[0]) {
                     res.send(rows);
                 }
-                connection.release();
             });
         }
     });
@@ -59,6 +61,7 @@ router.post('/getClassInfo', function(req, res) {
         }
         connection.query('SELECT * FROM Schedule_ WHERE CourseNumber = ?', [req.body.class], function(err2, rows) {
             if(err2) {
+              connection.release();
                 console.log('Error performing query: ' + err2);
                 throw err2;
             } else if (!rows[0]) {
@@ -68,6 +71,7 @@ router.post('/getClassInfo', function(req, res) {
                 schedID = rows[0].ScheduleID
                 connection.query('SELECT * FROM Enrollment WHERE ScheduleID = ?', [rows[0].ScheduleID], function(err3, rows) {
                     if(err3) {
+                      connection.release();
                         console.log('Error performing query: ' + err3);
                         throw err3;
                     } else if (rows[0]) {
@@ -78,6 +82,7 @@ router.post('/getClassInfo', function(req, res) {
                 });
                 connection.query('SELECT * FROM Placement WHERE ScheduleID = ?', schedID, function(err4, rows) {
                     if(err4) {
+                      connection.release();
                         console.log('Error performing query: ' + err4);
                         throw err4;
                     } else if (rows[0]) {
@@ -87,6 +92,7 @@ router.post('/getClassInfo', function(req, res) {
                     }
                 });
                 connection.query('SELECT * FROM Student_Request WHERE ScheduleID = ?', schedID, function(err5, rows) {
+                  connection.release();
                     if(err5) {
                         console.log('Error performing query: ' + err5);
                         throw err5;
@@ -95,7 +101,6 @@ router.post('/getClassInfo', function(req, res) {
                     } else if (!rows[0]) {
                         sendData.push(null);
                     }
-                    connection.release();
                     sendRes(res, sendData);
                 });
             }
@@ -118,6 +123,7 @@ router.post('/getStudentNameHours', function(req, res) {
         } else {
             connection.query('SELECT FirstName, LastName FROM User_ WHERE ASURITE_ID = ?', [req.body.studentID] , function(err2, rows) {
                 if(err2) {
+                  connection.release();
                     console.log('Error performing query: ' + err2);
                     throw err2;
                 } else if (rows[0]) {
@@ -126,6 +132,7 @@ router.post('/getStudentNameHours', function(req, res) {
             });
             connection.query('SELECT TimeCommitment FROM Application WHERE ASURITE_ID = ?', [req.body.studentID] , function(err3, rows) {
                 if(err3) {
+                  connection.release();
                     console.log('Error performing query: ' + err3);
                     throw err3;
                 } else if (rows[0]) {
@@ -134,6 +141,7 @@ router.post('/getStudentNameHours', function(req, res) {
             });
             connection.query('SELECT TAHours FROM Placement WHERE TA = ?', [req.body.studentID] , function(err4, rows) {
                 if(err4) {
+                  connection.release();
                     console.log('Error performing query: ' + err4);
                     throw err4;
                 } else if (rows[0]) {
@@ -147,6 +155,7 @@ router.post('/getStudentNameHours', function(req, res) {
             });
             connection.query('SELECT TATwoHours FROM Placement WHERE TATwo = ?', [req.body.studentID] , function(err5, rows) {
                 if(err5) {
+                  connection.release();
                     console.log('Error performing query: ' + err5);
                     throw err5;
                 } else if (rows[0]) {
@@ -159,6 +168,7 @@ router.post('/getStudentNameHours', function(req, res) {
             });
             connection.query('SELECT GraderOneHours FROM Placement WHERE GraderOne = ?', [req.body.studentID] , function(err5, rows) {
                 if(err5) {
+                  connection.release();
                     console.log('Error performing query: ' + err5);
                     throw err5;
                 } else if (rows[0]) {
@@ -170,6 +180,7 @@ router.post('/getStudentNameHours', function(req, res) {
                 }
             });
             connection.query('SELECT GraderTwoHours FROM Placement WHERE GraderTwo = ?', [req.body.studentID] , function(err5, rows) {
+              connection.release();
                 if(err5) {
                     console.log('Error performing query: ' + err5);
                     throw err5;
@@ -180,7 +191,6 @@ router.post('/getStudentNameHours', function(req, res) {
                     'GraderTwoHours' : 0}];
                     sendData.push(graderTwoHours);
                 }
-                connection.release();
                 sendRes(res, sendData);
             });
         }
@@ -196,17 +206,15 @@ function getAvailableHours(rows) {
   }
   console.log(asuID);
   var resultMap = {};
-  console.log("Gonna get Thread-Pool");
     mysql_pool.getConnection(function(err, connection) {
         if (err) {
             connection.release();
             console.log('Error getting mysql_pool connection: ' + err);
             throw err;
         } else {
-          console.log("before query");
             connection.query('SELECT TimeCommitment,ASURITE_ID FROM Application WHERE ASURITE_ID IN (?)', [asuID] , function(err3, rows) {
-              console.log("After query");
                 if(err3) {
+                  connection.release();
                     console.log('Error performing TimeCommitment query: ' + err3);
                     throw err3;
                 } else if (rows[0]) {
@@ -221,22 +229,20 @@ function getAvailableHours(rows) {
                       resultMap[rows[i].ASURITE_ID] = 0;
                     }
                   }
-                  console.log("HelloWorld::"+JSON.stringify(resultMap));
                     //sendData.push(rows);
                 }
             });
-            console.log("before query");
             connection.query('SELECT TAHours,TA FROM Placement WHERE TA IN (?)', [asuID] , function(err4, rows) {
-              console.log("After query");
                 if(err4) {
+                  connection.release();
                     console.log('Error performing TAHours query: ' + err4);
                     throw err4;
                 } else if (rows[0]) {
                     //sendData.push(rows);
-                    console.log(JSON.stringify(rows));
+                    //console.log(JSON.stringify(rows));
                     for(var i = 0; i<rows.length; i++)
                     {
-                      console.log("gonna loop"+rows[i].TA+"::"+resultMap[rows[i].TA]);
+                      //console.log("gonna loop"+rows[i].TA+"::"+resultMap[rows[i].TA]);
                       if(rows[i].TAHours != null && resultMap[rows[i].TA] != 0)
                       {
                         resultMap[rows[i].TA] = resultMap[rows[i].TA] - rows[i].TAHours;
@@ -244,10 +250,9 @@ function getAvailableHours(rows) {
                     }
                 }
             });
-            console.log("Before query");
             connection.query('SELECT TATwoHours , TATwo FROM Placement WHERE TATwo IN (?)', [asuID] , function(err5, rows) {
-              console.log("After query");
                 if(err5) {
+                  connection.release();
                     console.log('Error TATwoHours performing query: ' + err5);
                     throw err5;
                 } else if (rows[0])
@@ -261,11 +266,9 @@ function getAvailableHours(rows) {
                   }
                 }
             });
-            console.log("Before query");
             connection.query('SELECT GraderOneHours,GraderOne FROM Placement WHERE GraderOne IN (?) ', [asuID] , function(err5, rows) {
-              console.log("After query");
                 if(err5) {
-                    console.log('Error performing GraderOne query: ' + err5);
+                  connection.release();
                     throw err5;
                 } else if (rows[0]) {
                     for(var i = 0; i<rows.length; i++)
@@ -277,10 +280,8 @@ function getAvailableHours(rows) {
                     }
                 }
             });
-            console.log("Before query");
             connection.query('SELECT GraderTwoHours,GraderTwo FROM Placement WHERE GraderTwo IN (?)', [asuID] , function(err5, rows) {
               connection.release();
-              console.log("After query");
                 if(err5) {
                     console.log('Error performing GraderTwo query: ' + err5);
                     throw err5;
@@ -293,10 +294,8 @@ function getAvailableHours(rows) {
                     }
                   }
                 }
-                console.log("before return"+JSON.stringify(resultMap));
                 //return resultMap;
                 var availableHours = JSON.stringify(resultMap);
-                console.log(availableHours.length);
                 if(availableHours.length > 0)
                 {
                   resolve(availableHours);
@@ -320,12 +319,12 @@ router.post('/updateStatus', function(req, res) {
         } else {
             connection.query('UPDATE Schedule_ SET AssignedStatus = ? WHERE CourseNumber = ?', [req.body.status, req.body.class] , function(err2, rows) {
                 if(err2) {
+                  connection.release();
                     console.log('Error performing query: ' + err2);
                     throw err2;
                 } else {
                     res.sendStatus(200);
                 }
-                connection.release();
             });
         }
     });
@@ -344,16 +343,17 @@ router.post('/updateEnrollment', function(req, res) {
         } else {
             connection.query('SELECT * FROM Schedule_ WHERE CourseNumber = ?', [req.body.class], function(err2, rows) {
                 if(err2) {
+                  connection.release();
                     console.log('Error performing query: ' + err2);
                     throw err2;
                 } else if (rows[0]) {
                     schedID = rows[0].ScheduleID;
                     connection.query('INSERT INTO Enrollment (EnrollmentNumCurrent, DateEntered, ScheduleID) VALUES (?, ?, ?)', [req.body.enrollment, dateObj, schedID], function(err4, rows) {
+                      connection.release();
                         if(err4) {
                             console.log('Error performing query: ' + err4);
                             throw err4;
                         } else {
-                            connection.release();
                             res.sendStatus(200);
                         }
                     });
@@ -372,6 +372,7 @@ router.post('/updateRequiredHours', function(req, res) {
             throw err;
         } else {
             connection.query('UPDATE Schedule_ SET TARequiredHours = ?, GraderRequiredHours = ? WHERE CourseNumber = ?', [req.body.taHours, req.body.graderHours, req.body.class] , function(err2, rows) {
+              connection.release();
                 if(err2) {
                     console.log('Error performing query: ' + err2);
                     throw err2;
@@ -379,7 +380,6 @@ router.post('/updateRequiredHours', function(req, res) {
                     var sendData = [req.body.taHours, req.body.graderHours];
                     res.send(sendData);
                 }
-                connection.release();
             });
         }
     });
@@ -402,6 +402,7 @@ router.post('/updateAssignedStudents', function(req, res) {
         } else {
             connection.query('SELECT * FROM Schedule_ WHERE CourseNumber = ?', [classNumber], function(err2, rows) {
                 if(err2) {
+                  connection.release();
                     console.log('Error performing query: ' + err2);
                     throw err2;
                 } else if (rows[0]) {
@@ -409,6 +410,7 @@ router.post('/updateAssignedStudents', function(req, res) {
                     if (req.body.deleteClass) {
                         connection.query('DELETE FROM Placement WHERE ScheduleID = ?', [schedID], function(err3, rows) {
                             if(err3) {
+                              connection.release();
                                 console.log('Error performing query: ' + err3);
                                 throw err3;
                             } else {
@@ -418,6 +420,7 @@ router.post('/updateAssignedStudents', function(req, res) {
                     } else if (req.body.deleteStudent && req.body.deleteStudent === 1) {
                         connection.query('UPDATE Placement SET TA = null, TAStatus = null, TAHours = null WHERE ScheduleID = ?', [schedID], function(err4, rows) {
                             if(err4) {
+                              connection.release();
                                 console.log('Error performing query: ' + err4);
                                 throw err4;
                             } else {
@@ -428,6 +431,7 @@ router.post('/updateAssignedStudents', function(req, res) {
                     } else if (req.body.deleteStudent && req.body.deleteStudent === 2) {
                         connection.query('UPDATE Placement SET TATwo = null, TATwoStatus = null, TATwoHours = null WHERE ScheduleID = ?', [schedID], function(err5, rows) {
                             if(err5) {
+                              connection.release();
                                 console.log('Error performing query: ' + err5);
                                 throw err5;
                             } else {
@@ -438,6 +442,7 @@ router.post('/updateAssignedStudents', function(req, res) {
                     } else if (req.body.deleteStudent && req.body.deleteStudent === 3) {
                         connection.query('UPDATE Placement SET GraderOne = null, GraderOneStatus = null, GraderOneHours = null WHERE ScheduleID = ?', [schedID], function(err6, rows) {
                             if(err6) {
+                              connection.release();
                                 console.log('Error performing query: ' + err6);
                                 throw err6;
                             } else {
@@ -448,6 +453,7 @@ router.post('/updateAssignedStudents', function(req, res) {
                     } else if (req.body.deleteStudent && req.body.deleteStudent === 4) {
                         connection.query('UPDATE Placement SET GraderTwo = null, GraderTwoStatus = null, GraderTwoHours = null WHERE ScheduleID = ?', [schedID], function(err7, rows) {
                             if(err7) {
+                              connection.release();
                                 console.log('Error performing query: ' + err7);
                                 throw err7;
                             } else {
@@ -457,6 +463,7 @@ router.post('/updateAssignedStudents', function(req, res) {
                         });
                     } else {
                         connection.query('UPDATE Placement SET TA = ?, TAStatus = ?, TATwo = ?, TATwoStatus = ?, GraderOne = ?, GraderOneStatus = ?, GraderTwo = ?, GraderTwoStatus = ?, TAHours = ?, TATwoHours = ?, GraderOneHours = ?, GraderTwoHours = ? WHERE ScheduleID = ?', [req.body.ta, req.body.taStatus, req.body.taTwo, req.body.taTwoStatus, req.body.graderOne, req.body.graderOneStatus, req.body.graderTwo, req.body.graderTwoStatus, req.body.taAssignedHours, req.body.taTwoAssignedHours, req.body.graderOneAssignedHours, req.body.graderTwoAssignedHours, schedID], function(err8, rows) {
+                          connection.release();
                             if(err8) {
                                 console.log('Error performing query: ' + err8);
                                 throw err8;
@@ -465,7 +472,6 @@ router.post('/updateAssignedStudents', function(req, res) {
                             }
                         });
                     }
-                    connection.release();
                 }
             });
         }
@@ -501,6 +507,7 @@ router.post('/saveNewStudents', function(req, res) {
                 } else {
                     connection.query('SELECT ScheduleID FROM Schedule_ WHERE CourseNumber = ?', [req.body[0].class], function(err2, rows) {
                         if(err2) {
+                          connection.release();
                             console.log('Error performing query 2: ' + err2);
                             throw err2;
                         } else {
@@ -508,11 +515,13 @@ router.post('/saveNewStudents', function(req, res) {
                             connection.query('SELECT * FROM Placement WHERE ScheduleID = ?', [schedID], function(err3, rows) {
                                 //console.log("ROWS", rows)
                                 if(err3) {
+                                  connection.release();
                                     console.log('Error performing query 3: ' + err3);
                                     throw err3;
                                 } else if (rows.length > 0) {
                                     connection.query('UPDATE Placement SET ? WHERE ScheduleID = ?', [allStudents, schedID], function(err4, rows) {
                                         if(err4) {
+                                          connection.release();
                                             console.log('Error performing query 4: ' + err4);
                                             throw err4;
                                         } else {
@@ -522,6 +531,7 @@ router.post('/saveNewStudents', function(req, res) {
                                 } else {
                                     allStudents.ScheduleID = schedID;
                                     connection.query('INSERT INTO Placement SET ?', [allStudents], function(err5, rows) {
+                                      connection.release();
                                         if(err5) {
                                             console.log('Error performing query 5: ' + err5);
                                             throw err5;
@@ -530,7 +540,6 @@ router.post('/saveNewStudents', function(req, res) {
                                         }
                                     });
                                 }
-                                connection.release();
                             });
                         }
 
@@ -550,12 +559,12 @@ router.get('/getDeadline', function(req, res) {
             throw err;
         } else {
             connection.query('SELECT CurrentSemester, DeadlineDate FROM Deadline', function(err2, rows) {
+              connection.release();
                 if(err2) {
                     console.log('Error performing query: ' + err2);
                     throw err2;
                 }
                 res.send({date:rows[0].DeadlineDate, semester:rows[0].CurrentSemester});
-                connection.release();
             });
         }
     });
@@ -570,12 +579,12 @@ router.post('/setDeadline', function(req, res) {
             throw err;
         } else {
             connection.query('UPDATE Deadline SET DeadlineDate = ?, CurrentSemester = ?', [req.body.date, req.body.semester], function(err2, rows) {
+              connection.release();
                 if(err2) {
                     console.log('Error performing query: ' + err2);
                     throw err2;
                 }
                 res.sendStatus(200);
-                connection.release();
             });
         }
     });
@@ -590,6 +599,7 @@ function checkForNull() {
             throw err;
         } else {
             connection.query('DELETE FROM Placement WHERE (TA IS null AND TATwo IS null AND GraderOne IS null AND GraderTwo IS null)', function(err2, rows) {
+              connection.release();
                 if(err2) {
                     console.log('Error performing query: ' + err2);
                     throw err2;
@@ -610,6 +620,7 @@ router.get('/evaluations/appnames', function(req, res) {
             throw err;
         }
         connection.query("SELECT CONCAT(FirstName, ' ', LastName) AS StudentName, ASURITE_ID FROM User_ WHERE UserRole = 'student'", function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -628,7 +639,6 @@ router.get('/evaluations/appnames', function(req, res) {
                 var response = JSON.stringify(Names);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -643,21 +653,18 @@ router.post('/evaluationsPC', function(req, res) {
         }
         connection.query('INSERT INTO Student_Evaluation SET ?', [req.body], function(err2) {
             if(err2) {
+              connection.release();
                 console.log('Error performing query: ' + err2);
                 throw err2;
             } else {
                 res.sendStatus(200);
                 }
         });
-        connection.release();
     });
 });
 
 // Get all Students in Database
 router.post('/applicationNames', function(req, res) {
-  console.log("application Name called");
-  //console.log("Thread-Pool size::"+process.env.UV_THREADPOOL_SIZE);
-  console.log("gonna get Thread-Pool");
     mysql_pool.getConnection(function(err, connection) {
         if (err) {
             connection.release();
@@ -680,10 +687,8 @@ router.post('/applicationNames', function(req, res) {
             query = "SELECT concat(FirstName, ' ', LastName) Name, AppStatus, User_.ASURITE_ID, Application.rating From User_ INNER JOIN Application ON User_.ASURITE_ID = Application.ASURITE_ID WHERE UserRole = 'student' ORDER BY Name DESC";
           }
         }
-        console.log("Gonna Query");
         connection.query(query, [req.body.course], function(err2, rows) {
           connection.release();
-          console.log("After Query");
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -695,7 +700,7 @@ router.post('/applicationNames', function(req, res) {
                 //console.log("application Names::"+jsonObject);
                   var Students = [];
                   for (var i = 0; i < rows.length; i++) {
-                    console.log(jsonObject[rows[i].ASURITE_ID]);
+                    //console.log(jsonObject[rows[i].ASURITE_ID]);
                        var FullName = {
                       'Name'       : rows[i].Name,
                       'AppStatus'  : rows[i].AppStatus,
@@ -705,7 +710,7 @@ router.post('/applicationNames', function(req, res) {
                       Students.push(FullName);
                   }
                   var response = JSON.stringify(Students);
-                  console.log("response::"+response);
+                  //console.log("response::"+response);
                   res.send(response);
               },function(error){
                  console.log('An error in promise occurred: ' + error)
@@ -720,7 +725,7 @@ router.post('/applicationNames', function(req, res) {
 // Get Student Applications
 router.post('/applications', function(req, res) {
     var queryString = '';
-    console.log(req.body);
+    //console.log(req.body);
     if (!req.body.filter || req.body.filter === 1) {
         queryString = "SELECT concat(FirstName, ' ', LastName) Name, isQualified, isPrefer, isPreviouslyTA, isPreviouslyGrader, GPA, isFullTime, User_.ASURITE_ID, Application.rating From User_ INNER JOIN Course_Competencies ON User_.ASURITE_ID = Course_Competencies.ASURITE_ID INNER JOIN Application ON User_.ASURITE_ID = Application.ASURITE_ID WHERE isCourse = ? AND (isAcademicProbation = 0 OR isAcademicProbation IS NULL) AND AppStatus = 'complete' AND (SpeakTest >= 26 OR SpeakTest IS NULL) AND ((isQualified = 1 OR isPrefer = 1 OR isQualified = 0 OR isPrefer = 0) OR (isPreviouslyGrader = 1 OR isPreviouslyTA = 1 OR isPreviouslyGrader = 0 OR isPreviouslyTA = 0)) ORDER BY isQualified DESC, isPrefer DESC, isPreviouslyTA DESC, isPreviouslyGrader DESC, GPA DESC, isFullTime DESC, Name DESC";
     } else if (req.body.filter === 2) {
@@ -737,16 +742,14 @@ router.post('/applications', function(req, res) {
     else if (req.body.filter === 7) {
         queryString = "SELECT concat(FirstName, ' ', LastName) Name, isQualified, isPrefer, isPreviouslyTA, isPreviouslyGrader, GPA, isFullTime, User_.ASURITE_ID, Application.rating From User_ INNER JOIN Course_Competencies ON User_.ASURITE_ID = Course_Competencies.ASURITE_ID INNER JOIN Application ON User_.ASURITE_ID = Application.ASURITE_ID WHERE isCourse = ? AND (isAcademicProbation = 0 OR isAcademicProbation IS NULL) AND AppStatus = 'complete' AND (SpeakTest >= 26 OR SpeakTest IS NULL) AND ((isQualified = 1 OR isPrefer = 1) OR (isPreviouslyGrader = 1 OR isPreviouslyTA = 1)) ORDER BY Application.rating DESC , isFullTime DESC, isQualified DESC, isPrefer DESC, isPreviouslyTA DESC, isPreviouslyGrader DESC, GPA DESC, Name DESC";
     }
-    console.log("Gonna take mysql connection");
     mysql_pool.getConnection(function(err, connection) {
         if (err) {
             connection.release();
             console.log('Error getting mysql_pool connection: ' + err);
             throw err;
         }
-        console.log("Gonna Query");
         connection.query(queryString, [req.body.course], function(err2, rows) {
-          console.log("After query");
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -770,7 +773,6 @@ router.post('/applications', function(req, res) {
                 var response = JSON.stringify(StudentNames);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -784,6 +786,7 @@ router.post('/contactInfo', function(req, res) {
             throw err;
         }
         connection.query("SELECT CONCAT(User_.FirstName, ' ', User_.LastName) AS Name, UserEmail, PhoneNumber, MobileNumber, AddressOne, AddressTwo, AddressCountry, AddressCity, AddressState, AddressZip FROM Application INNER JOIN User_ ON Application.ASURITE_ID = User_.ASURITE_ID WHERE Application.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -805,7 +808,6 @@ router.post('/contactInfo', function(req, res) {
                 var response = JSON.stringify(ContactInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -819,6 +821,7 @@ router.post('/applicationTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT AppID, EducationLevel, DegreeProgram, isFourPlusOne, GPA, isInternationalStudent, SpeakTest, DATE_FORMAT(FirstSession, '%m/%d/%Y') AS FirstSession, GraduationDate, TimeCommitment, isTA, isGrader, CurrentEmployer, WorkHours, isWorkedASU FROM Application INNER JOIN User_ ON Application.ASURITE_ID = User_.ASURITE_ID WHERE Application.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -845,7 +848,6 @@ router.post('/applicationTable', function(req, res) {
                 var response = JSON.stringify(AppInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -859,7 +861,7 @@ router.post('/override', function(request, res) {
         }
         var query;
         var updateValue;
-        console.log(request.body);
+        //console.log(request.body);
         if(request.body.gpa)
         {
           query = "UPDATE Application set GPA = ? where AppID = ?";
@@ -867,12 +869,12 @@ router.post('/override', function(request, res) {
         }
         else
         {
-            console.log("Inside four-plus-one");
             query = "UPDATE Application set isFourPlusOne = ? where AppID = ?";
             updateValue = request.body.fourPlusOne;
         }
 
         connection.query(query, [updateValue,request.body.appID], function(err2) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -881,7 +883,6 @@ router.post('/override', function(request, res) {
                 }
         });
         //connection.end();
-        connection.release();
     });
 });
 
@@ -894,6 +895,7 @@ router.post('/languagesTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT isLanguage, LanguageLevel, OtherLanguage, OtherLevel FROM Languages INNER JOIN User_ ON Languages.ASURITE_ID = User_.ASURITE_ID WHERE Languages.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -912,7 +914,6 @@ router.post('/languagesTable', function(req, res) {
                 var response = JSON.stringify(LanguageInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -926,6 +927,7 @@ router.post('/ideTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT isIDE, OtherIDE FROM IDEs INNER JOIN User_ ON IDEs.ASURITE_ID = User_.ASURITE_ID WHERE IDEs.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -942,7 +944,6 @@ router.post('/ideTable', function(req, res) {
                 var response = JSON.stringify(IDEInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -956,6 +957,7 @@ router.post('/toolsTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT isTool, OtherTool FROM Collaborative_Tools INNER JOIN User_ ON Collaborative_Tools.ASURITE_ID = User_.ASURITE_ID WHERE Collaborative_Tools.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -972,7 +974,6 @@ router.post('/toolsTable', function(req, res) {
                 var response = JSON.stringify(ToolInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -986,6 +987,7 @@ router.post('/coursesTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT isCourse, isPrefer, isQualified, isPreviouslyTA, isPreviouslyGrader, OtherCourse, isOtherPrefer, isOtherQualified, isOtherPreviouslyTA, isOtherPreviouslyGrader FROM Course_Competencies INNER JOIN User_ ON Course_Competencies.ASURITE_ID = User_.ASURITE_ID WHERE Course_Competencies.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -1010,7 +1012,6 @@ router.post('/coursesTable', function(req, res) {
                 var response = JSON.stringify(CourseInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -1024,6 +1025,7 @@ router.post('/calendarTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT CalendarDay, TIME_FORMAT(StartHour, '%h:%i %p') AS StartHour, TIME_FORMAT(StopHour, '%h:%i %p') AS StopHour FROM Calendar INNER JOIN User_ ON Calendar.ASURITE_ID = User_.ASURITE_ID WHERE Calendar.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -1041,7 +1043,6 @@ router.post('/calendarTable', function(req, res) {
                 var response = JSON.stringify(CalendarInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -1055,6 +1056,7 @@ router.post('/attachmentTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT IposName, TranscriptName, ResumeName, DATE_FORMAT(IposUploadDate, '%m/%d/%Y') AS IposUploadDate, DATE_FORMAT(TranscriptUploadDate, '%m/%d/%Y') AS TranscriptUploadDate, DATE_FORMAT(ResumeUploadDate, '%m/%d/%Y') AS ResumeUploadDate, Attachment.ASURITE_ID FROM Attachment INNER JOIN User_ ON Attachment.ASURITE_ID = User_.ASURITE_ID WHERE Attachment.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -1073,7 +1075,6 @@ router.post('/attachmentTable', function(req, res) {
                 var response = JSON.stringify(AttachmentInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -1114,6 +1115,7 @@ router.post('/evaluationTable', function(req, res) {
             throw err;
         }
         connection.query("SELECT DATE_FORMAT(DateCreated, '%m/%d/%Y') AS DateCreated, InstructorName, QOneScore, QOneComments, QTwoScore, QTwoComments, QThreeScore, QThreeComments, QFourScore, QFourComments FROM Student_Evaluation WHERE Student_Evaluation.ASURITE_ID = ?", [req.body.studentId], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -1138,7 +1140,6 @@ router.post('/evaluationTable', function(req, res) {
                 var response = JSON.stringify(EvalInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -1152,6 +1153,7 @@ router.post('/getFeedback', function(req, res) {
             throw err;
       }
       connection.query('SELECT FeedbackText, hasComplied FROM Feedback WHERE AppID = ?', [req.body.appID], function(err2, rows) {
+        connection.release();
           if(err2) {
             console.log('Error performing query: ' + err2);
             throw err2;
@@ -1174,12 +1176,14 @@ router.post('/saveFeedback', function(req, res) {
       }
       connection.query('SELECT FeedbackID FROM Feedback WHERE AppID = ?', [req.body.appID], function(err2, rows) {
           if(err2) {
+            connection.release();
             console.log('Error performing query: ' + err2);
             throw err2;
           }
           if (!rows.length) {
               connection.query('INSERT INTO Feedback (FeedbackText, hasComplied, AppID) VALUES (?, ?, ?)', [req.body.feedback, req.body.hasComplied, req.body.appID], function(err3, rows) {
                   if(err3) {
+                    connection.release();
                     console.log('Error performing query: ' + err3);
                     throw err3;
                   }
@@ -1188,11 +1192,11 @@ router.post('/saveFeedback', function(req, res) {
               });
           } else if (rows[0]) {
               connection.query('UPDATE Feedback SET FeedbackText = ?, hasComplied = ? WHERE AppID = ?', [req.body.feedback, req.body.hasComplied, req.body.appID], function(err4, rows) {
+                connection.release();
                   if(err4) {
                     console.log('Error performing query: ' + err4);
                     throw err4;
                   }
-                  connection.release();
                   res.sendStatus(200);
               });
           }
@@ -1209,11 +1213,11 @@ router.post('/removeFeedback', function(req, res) {
             throw err;
       }
       connection.query('DELETE FROM Feedback WHERE AppID = ?', [req.body.appID], function(err2, rows) {
+        connection.release();
           if(err2) {
             console.log('Error performing query: ' + err2);
             throw err2;
           }
-          connection.release();
           res.sendStatus(200);
       });
    });
@@ -1228,11 +1232,11 @@ router.post('/ApplicationRating', function(req, res) {
             throw err;
       }
       connection.query('UPDATE Application set rating = ? WHERE AppID = ?', [req.body.rating,req.body.appID], function(err2, rows) {
+        connection.release();
           if(err2) {
             console.log('Error performing query: ' + err2);
             throw err2;
           }
-          connection.release();
           res.sendStatus(200);
       });
    });
@@ -1247,6 +1251,7 @@ router.get('/courseEdit', function(req, res) {
             throw err;
         }
         connection.query("SELECT CourseSection, CourseName FROM Courses", function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -1269,7 +1274,6 @@ router.get('/courseEdit', function(req, res) {
                 var response = JSON.stringify(CourseInfo);
                 res.send(response);
             }
-            connection.release();
         });
     });
 });
@@ -1284,6 +1288,7 @@ router.post('/pcSetUserPassword', function(req, res) {
             throw err;
         }
         connection.query('SELECT UserPassword FROM User_ WHERE ASURITE_ID = ?', [req.body.ASURITE_ID], function(err2, rows){
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -1296,7 +1301,6 @@ router.post('/pcSetUserPassword', function(req, res) {
             } else {
                 res.send({'error' : 1}); // Responds error 1 if user not found
             }
-            connection.release();
         });
     });
 });
@@ -1320,13 +1324,13 @@ function changePassword(response, req, validation) {
                         throw err;
                     }
                     connection.query('UPDATE User_ SET UserPassword = ? WHERE ASURITE_ID = ?', [req.body.NewPassword, req.body.ASURITE_ID], function(err2, rows) {
+                      connection.release();
                         if(err2) {
                             console.log('Error performing query: ' + err2);
                             throw err2;
                         } else {
                             response.sendStatus(200);
                         }
-                        connection.release();
                     });
                 });
             }
@@ -1346,13 +1350,13 @@ router.post('/updateCourses', function(req, res) {
             throw err;
         }
         connection.query("UPDATE Courses SET CourseSection = ?, CourseName = ? WHERE CourseSection = ?", [req.body.newCourseSection, req.body.courseName, req.body.courseSection], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
             } else {
                 res.sendStatus(200);
             }
-            connection.release();
         });
     });
 });
@@ -1366,13 +1370,13 @@ router.post('/deleteCourses', function(req, res) {
             throw err;
         }
         connection.query("DELETE FROM Courses WHERE CourseSection = ?", [req.body.CourseSection], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
             } else {
                 res.sendStatus(200);
             }
-            connection.release();
         });
     });
 });
@@ -1386,13 +1390,13 @@ router.post('/addCourses', function(req, res) {
             throw err;
         }
         connection.query("INSERT INTO Courses (CourseSection, CourseName) VALUES (?,?)", [req.body.courseAddSection, req.body.courseAddName], function(err2, rows) {
+          connection.release();
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
             } else {
                 res.sendStatus(200);
             }
-            connection.release();
         });
     });
 });
@@ -1400,18 +1404,14 @@ router.post('/addCourses', function(req, res) {
 // Gets Courses from Courses Table
 router.get('/courses', function(req, res) {
 
-  console.log("courses called");
-
     mysql_pool.getConnection(function(err, connection) {
         if (err) {
             connection.release();
             console.log('Error getting mysql_pool connection: ' + err);
             throw err;
         }
-        console.log("courses before query");
         connection.query("SELECT CourseSection FROM Courses", function(err2, rows) {
           connection.release();
-          console.log("courses after query");
             if(err2) {
                 console.log('Error performing query: ' + err2);
                 throw err2;
@@ -1448,13 +1448,13 @@ router.post('/editSchedule', function(req, res) {
             throw err;
         } else {
             connection.query('UPDATE Schedule_ SET SessionIs = ?, Location = ?, Subject = ?, CatalogNumber = ?, CourseTitle = ?, FirstName = ?, LastName = ?, EnrollmentNumPrev = ? WHERE CourseNumber = ?', [req.body.session, req.body.location, req.body.subject, req.body.catalog, req.body.title, req.body.firstName, req.body.lastName, req.body.enrollment, req.body.course] , function(err2, rows) {
+              connection.release();
                 if(err2) {
                     console.log('Error performing query: ' + err2);
                     throw err2;
                 } else {
                     res.sendStatus(200);
                 }
-                connection.release();
             });
         }
     });
